Submission comments
Use two npm repositories to reduce amount of code that needs to be written
and tested. csvtojson to convert csv to json and count-lines-in-file to
count the number of lines in the input csv file.

Use csvtojson to create an array of json objects and then user
JSON.stringify to write the array to a file.

Tested by running:
npm install -y
node csv_to_json.js

and diffing the output with the provided solution. I am aware that the
line endings differ but I see no need to go out of my way to add
dos line endings.